fix: use project-id/tenant-id instead of underscore variants in miggo-collector
We still have a few components ingesting metrcis/traces with
`project-id`/`tenant-id` as attributres, instead of
`project_id`/`tenant_id`. This is causing errors with the prometheus
exporter, since labels there are normalized. It will replace all `-`
with `_`, which means we end up with duplicated labels.

The fix is to simply always use the dashed variants instead of the
underscore.

Towards: MIG-7788